split main.dart into smaller files
